@types/highcharts-ng - Update for highcharts v6
