+++ types/highcharts-ng/tslint.json
@@ -1,6 +1,6 @@
 {
-    "extends": "dtslint/dt.json",
-    "rules": {
+    "extends": "dtslint/dt.json"
+    /*"rules": {
         "adjacent-overload-signatures": false,
         "array-type": false,
         "arrow-return-shorthand": false,
@@ -75,5 +75,5 @@
         "unified-signatures": false,
         "void-return": false,
         "whitespace": false
-    }
+    }*/
 }
